refactor(*): reconstruct Migration structure
