+++ migration.md
@@ -16,6 +16,8 @@ extension Defaults.Keys {
 1. Call `Defaults.migration(.arrayString)`, `Defaults.migration(.dictionaryStringInt)`.
 2. Now `Defaults[.arrayString]`, `Defaults[.dictionaryStringInt]` should be readable.
 
+---
+
 ### From `Codable Array/Dictionary` to `Native Array/Dictionary/Set`(With Codable Elements)
 
 Before v4.X, `Defaults` will store array/dictionary as a json string.  
@@ -44,6 +46,7 @@ private struct TimeZone: Defaults.Serializable & Codable {
 2. Call `Defaults.migration(.arrayTimezone)`, `Defaults.migration(.dictionaryTimezone)`.
 3. Now `Defaults[.arrayTimezone]`, `Defaults[.dictionaryTimezone]` should be readable.
 
+---
 
 ### From `Codable` struct to `Dictionary` (Optional) 
 
@@ -91,7 +94,7 @@ private struct TimeZoneBridge: Defaults.Bridge {
 	}
 }
 ```
-2. Let `TimeZone` protocol conform to `Defaults.NativeType` and its static bridge is `TimeZoneBridge`(Compiler will complain `TimeZone` is not conform to Defaults.NativeType, will resolve it later).
+2. Let `TimeZone` protocol conform to `Defaults.NativeType` and its static bridge is `TimeZoneBridge`(Compiler will complain `TimeZone` is not conform to `Defaults.NativeType`, will resolve it later).
 ```swift
 private struct TimeZone: Defaults.NativeType, Hashable {
